refactor: extract ServiceManager trait to reusable module
- Move ServiceManager trait, MockServiceManager, and SystemDServiceManager to new service_manager.rs module
- Update bluetooth page to import from service_manager module
- Migrate existing service manager tests to new module

Coding-Agent: Claude Code
Model: claude-sonnet-4-5
